refactor(routes): share logic between new route resources
