+++ pkg/plugins/policies/meshhttproute/plugin/v1alpha1/gateway_routes.go
@@ -1,7 +1,6 @@
 package v1alpha1
 
 import (
-	"context"
 	"slices"
 	"sort"
 	"strings"
@@ -10,143 +9,34 @@ import (
 
 	common_api "github.com/kumahq/kuma/api/common/v1alpha1"
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
-	"github.com/kumahq/kuma/pkg/core/permissions"
 	core_mesh "github.com/kumahq/kuma/pkg/core/resources/apis/mesh"
 	"github.com/kumahq/kuma/pkg/core/resources/model"
-	core_xds "github.com/kumahq/kuma/pkg/core/xds"
 	"github.com/kumahq/kuma/pkg/plugins/policies/core/rules"
+	"github.com/kumahq/kuma/pkg/plugins/policies/core/xds/meshroute"
 	api "github.com/kumahq/kuma/pkg/plugins/policies/meshhttproute/api/v1alpha1"
 	plugin_gateway "github.com/kumahq/kuma/pkg/plugins/runtime/gateway"
 	"github.com/kumahq/kuma/pkg/plugins/runtime/gateway/match"
 	"github.com/kumahq/kuma/pkg/plugins/runtime/gateway/route"
 	"github.com/kumahq/kuma/pkg/util/pointer"
 	xds_context "github.com/kumahq/kuma/pkg/xds/context"
-	envoy_names "github.com/kumahq/kuma/pkg/xds/envoy/names"
 	"github.com/kumahq/kuma/pkg/xds/envoy/tags"
-	xds_topology "github.com/kumahq/kuma/pkg/xds/topology"
 )
-
-type sublistener struct {
-	Hostname string
-	Tags     map[string]string
-	TLS      *mesh_proto.MeshGateway_TLS_Conf
-}
-type listenersHostnames struct {
-	listener     *mesh_proto.MeshGateway_Listener
-	sublisteners []sublistener
-}
-
-func CollectListenerInfos(
-	ctx context.Context,
-	meshCtx xds_context.MeshContext,
-	gateway *core_mesh.MeshGatewayResource,
-	proxy *core_xds.Proxy,
-	rawRules rules.GatewayRules,
-) []plugin_gateway.GatewayListenerInfo {
-	networking := proxy.Dataplane.Spec.GetNetworking()
-	listenersByPort := map[uint32]listenersHostnames{}
-	for _, listener := range gateway.Spec.GetConf().GetListeners() {
-		if listener.Protocol == mesh_proto.MeshGateway_Listener_TCP {
-			continue
-		}
-
-		listenerAcc, ok := listenersByPort[listener.GetPort()]
-		if !ok {
-			listenerAcc = listenersHostnames{
-				listener: listener,
-			}
-		}
-		hostname := listener.GetNonEmptyHostname()
-		listenerAcc.sublisteners = append(listenerAcc.sublisteners, sublistener{
-			Hostname: hostname,
-			Tags: mesh_proto.Merge(
-				networking.GetGateway().GetTags(),
-				gateway.Spec.GetTags(),
-				listener.GetTags(),
-			),
-			TLS: listener.GetTls(),
-		})
-		listenersByPort[listener.GetPort()] = listenerAcc
-	}
-
-	var infos []plugin_gateway.GatewayListenerInfo
-
-	for port, listener := range listenersByPort {
-		externalServices := meshCtx.Resources.ExternalServices()
-
-		matchedExternalServices := permissions.MatchExternalServicesTrafficPermissions(
-			proxy.Dataplane, externalServices, meshCtx.Resources.TrafficPermissions(),
-		)
-
-		outboundEndpoints := core_xds.EndpointMap{}
-		for k, v := range meshCtx.EndpointMap {
-			outboundEndpoints[k] = v
-		}
-
-		esEndpoints := xds_topology.BuildExternalServicesEndpointMap(
-			ctx,
-			meshCtx.Resource,
-			matchedExternalServices,
-			meshCtx.DataSourceLoader,
-			proxy.Zone,
-		)
-		for k, v := range esEndpoints {
-			outboundEndpoints[k] = v
-		}
-
-		hostInfos := SortRulesToHosts(
-			meshCtx.Resources.MeshLocalResources,
-			rawRules,
-			networking.Address,
-			listener.listener,
-			listener.sublisteners,
-		)
-		var filters []plugin_gateway.GatewayListenerFilter
-		for _, sublistener := range listener.sublisteners {
-			filters = append(filters, plugin_gateway.GatewayListenerFilter{
-				Hostnames: []string{sublistener.Hostname},
-				TLS:       sublistener.TLS,
-			})
-		}
-		infos = append(infos, plugin_gateway.GatewayListenerInfo{
-			Proxy:             proxy,
-			Gateway:           gateway,
-			HostInfos:         hostInfos,
-			ExternalServices:  externalServices,
-			OutboundEndpoints: outboundEndpoints,
-			Listener: plugin_gateway.GatewayListener{
-				Port:     port,
-				Protocol: listener.listener.GetProtocol(),
-				ResourceName: envoy_names.GetGatewayListenerName(
-					gateway.Meta.GetName(),
-					listener.listener.GetProtocol().String(),
-					port,
-				),
-				CrossMesh: listener.listener.GetCrossMesh(),
-				Resources: listener.listener.GetResources(),
-				Filters:   filters,
-			},
-		})
-	}
-
-	return infos
-}
 
 type ruleByHostname struct {
 	Rule     ToRouteRule
 	Hostname string
 }
 
-func SortRulesToHosts(
+func sortRulesToHosts(
 	meshLocalResources xds_context.ResourceMap,
 	rawRules rules.GatewayRules,
 	address string,
 	listener *mesh_proto.MeshGateway_Listener,
-	hostnameTags []sublistener,
+	sublisteners []meshroute.Sublistener,
 ) []plugin_gateway.GatewayHostInfo {
 	var hostInfos []plugin_gateway.GatewayHostInfo
 
-	for _, hostnameTag := range hostnameTags {
+	for _, hostnameTag := range sublisteners {
 		inboundListener := rules.NewInboundListenerHostname(
 			address,
 			listener.GetPort(),
