+++ pkg/plugins/policies/meshhttproute/plugin/v1alpha1/plugin.go
@@ -6,6 +6,7 @@ import (
 	"github.com/pkg/errors"
 
 	common_api "github.com/kumahq/kuma/api/common/v1alpha1"
+	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
 	core_plugins "github.com/kumahq/kuma/pkg/core/plugins"
 	core_mesh "github.com/kumahq/kuma/pkg/core/resources/apis/mesh"
 	core_model "github.com/kumahq/kuma/pkg/core/resources/model"
@@ -129,12 +130,14 @@ func ApplyToGateway(
 		return nil
 	}
 
-	listeners := CollectListenerInfos(
+	listeners := meshroute.CollectListenerInfos(
 		ctx,
 		xdsCtx.Mesh,
 		gateway,
 		proxy,
 		rawRules,
+		[]mesh_proto.MeshGateway_Listener_Protocol{mesh_proto.MeshGateway_Listener_HTTP, mesh_proto.MeshGateway_Listener_HTTPS},
+		sortRulesToHosts,
 	)
 	plugin_gateway.SetGatewayListeners(proxy, listeners)
